\name{mouse4302barcodevecs}
\docType{data}
\alias{mouse4302barcodevecs}
\title{mouse4302 data for barcode}
\description{Data used by the barcode package for microarrays of type mouse4302.}
\usage{
  data(bcparams-GPL1261)
}
\format{
  A matrix of parameters estimates used in the barcode function for
  Affymetrix mouse4302 microarray data.
  }
\keyword{datasets}


